backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
python-dotenv
pydantic[email]